rework tests to fit into a more standard pattern
