--- Makefile
@@ -1,4 +1,0 @@
-test:
-	node test.js
-
-.PHONY: test
